revamped the home page, added a footer and made the resume previews on the homepage show the correct preview, also added functionality to print and save resumes as pdf
