+++ package.json
@@ -35,6 +35,7 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.1",
     "react-hot-toast": "^2.5.2",
+    "react-to-print": "^3.1.0",
     "resend": "^4.5.0",
     "tailwind-merge": "^3.2.0",
     "tw-animate-css": "^1.2.8",
