+++ src/app/(main)/layout.jsx
@@ -1,11 +1,13 @@
 import React from "react";
 import Navbar from "./_components/Navbar";
+import Footer from "./_components/Footer";
 
 const Layout = ({ children }) => {
   return (
-    <div>
+    <div className="min-h-screen flex flex-col">
       <Navbar />
-      {children}
+      <main className="flex-grow">{children}</main>
+      <Footer />
     </div>
   );
 };
