+++ src/app/(main)/resumes/_components/AddNewResumeBtn.jsx
@@ -15,7 +15,11 @@ const AddNewResumeBtn = () => {
   };
 
   return (
-    <Button onClick={addNewResume} className="flex items-center gap-2" size={"lg"}>
+    <Button 
+      onClick={addNewResume} 
+      className="flex items-center gap-2 bg-blue-600 hover:bg-blue-700 text-white rounded-md px-4 py-2 transition-colors" 
+      size={"lg"}
+    >
       <Plus size={20} /> New Resume
     </Button>
   );
