+++ src/app/(main)/resumes/_components/ResumeItem.jsx
@@ -1,9 +1,10 @@
-import React from "react";
+import React, { useRef, useState } from "react";
 import ResumePreview from "@/components/ResumePreview";
 import axiosInstance from "@/lib/axios";
 import { cn, getLastUpdatedTime } from "@/lib/utils";
 import toast from "react-hot-toast";
 import { X, EllipsisVertical } from "lucide-react";
+import { useReactToPrint } from "react-to-print";
 
 const ResumeItem = ({
   resume,
@@ -12,7 +13,15 @@ const ResumeItem = ({
   toggleMenu,
   navigateToResume,
   setResumesState,
+  resumesState,
 }) => {
+  const [isDeleting, setIsDeleting] = useState(false);
+  const contentRef = useRef(null);
+
+  const reactToPrintFn = useReactToPrint({
+    contentRef,
+    documentTitle : resume?.data?.generalInfo?.title || `resume-${resume?.resumeId}`,
+  })
   return (
     <>
       <div className="w-full relative group">
@@ -34,19 +43,29 @@ const ResumeItem = ({
         </div>
 
         {openMenuId === resume._id && (
-          <div className="absolute top-8 right-2 bg-white shadow-lg rounded-md text-sm z-20 w-28 py-2">
+          <div className="absolute top-8 right-2 bg-white border border-gray-300 shadow-lg rounded-md text-sm z-20 w-28 py-2">
             <button
               onClick={() => {
                 setOpenMenuId(null);
-                // TODO: Implement print logic
+                reactToPrintFn()
               }}
-              className="w-full px-4 py-2 hover:bg-gray-100 text-left"
+              className="w-full px-4 py-2 hover:bg-gray-100 text-left border-b"
             >
               Print
             </button>
             <button
+              onClick={() => {
+                setOpenMenuId(null);
+                // TODO: Implement export logic
+              }}
+              className="w-full px-4 py-2 hover:bg-gray-100 text-left border-b"
+            >
+              Export
+            </button>
+            <button
               onClick={async () => {
                 try {
+                  setIsDeleting(true);
                   const response = await axiosInstance.delete(
                     "/resumes/delete",
                     {
@@ -71,11 +90,14 @@ const ResumeItem = ({
                   toast.error(
                     "Failed to delete resume, please try again later"
                   );
+                } finally {
+                  setIsDeleting(false);
                 }
               }}
               className="w-full px-4 py-2 hover:bg-gray-100 text-left text-red-500"
+              disabled={isDeleting}
             >
-              Delete
+              {isDeleting ? "Deleting..." : "Delete"}
             </button>
           </div>
         )}
@@ -86,14 +108,14 @@ const ResumeItem = ({
         >
           <div className="w-full aspect-[4/5] relative mb-2 sm:mb-3">
             {/* <Image
-                  src={placeholderImage}
-                  alt={`${resume?.data?.generalInfo?.title} thumbnail`}
-                  fill
-                  sizes="(max-width: 640px) 45vw, (max-width: 768px) 30vw, (max-width: 1024px) 25vw, 20vw"
-                  style={{ objectFit: "cover" }}
-                  className="rounded-md"
-                /> */}
-            <ResumePreview />
+                src={placeholderImage}
+                alt={`${resume?.data?.generalInfo?.title} thumbnail`}
+                fill
+                sizes="(max-width: 640px) 45vw, (max-width: 768px) 30vw, (max-width: 1024px) 25vw, 20vw"
+                style={{ objectFit: "cover" }}
+                className="rounded-md"
+              /> */}
+            <ResumePreview resumeData={resume.data} contentRef={contentRef} />
           </div>
           <h2 className="text-sm font-semibold truncate w-full">
             {resume?.data?.generalInfo?.title || "untitled"}
