+++ src/app/(main)/resumes/layout.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const layout = ({ children }) => {
   return (
-    <div className="h-full">
+    <div className="h-screen overflow-y-hidden">
       {children}
     </div>
   );
