+++ src/app/(main)/resumes/page.jsx
@@ -1,6 +1,9 @@
 import React from "react";
 import MyResume from "./_components/MyResume";
 import AddNewResumeBtn from "./_components/AddNewResumeBtn";
+import { LucideUpload } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import ImportResumeBtn from "./_components/ImportResumeBtn";
 
 export const metadata = {
   title: "Your Resumes",
@@ -8,16 +11,23 @@ export const metadata = {
 
 const Page = () => {
   return (
-    <div className="bg-muted min-h-[calc(100vh-64px)] p-3 sm:p-4 md:p-6 flex flex-col gap-6 md:gap-12">
+    <div className="bg-muted h-full p-3 sm:p-4 md:p-6 flex flex-col gap-6 md:gap-12 overflow-y-auto">
       <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-4 sm:gap-0">
         <div>
-          <h1 className="text-blue-950 font-bold text-xl sm:text-2xl">Resume Builder</h1>
+          <h1 className="text-blue-950 font-bold text-xl sm:text-2xl">
+            Resume Builder
+          </h1>
           <p className="text-gray-500 text-sm sm:text-base">
-            Create your own custom resume and land your dream job
+            Craft your professional resume and land your dream job.
           </p>
         </div>
-        <div className="w-full sm:w-auto">
-         <AddNewResumeBtn />
+        <div className="flex gap-4">
+          <div>
+          <ImportResumeBtn />
+          </div>
+          <div className="w-full sm:w-auto">
+            <AddNewResumeBtn />
+          </div>
         </div>
       </div>
       <MyResume />
