+++ src/app/globals.css
@@ -130,3 +130,20 @@
   -ms-overflow-style: none;  /* IE and Edge */
   scrollbar-width: none;     /* Firefox */
 }
+
+
+@media print {
+  #resumePreviewContent {
+    zoom: 1 !important;
+    padding: 0;
+  }
+
+  #resumePreviewContent {
+    font-family: Inter;
+  }
+
+  @page {
+    size: A4;
+    margin: 0.6cm;
+  }
+}
