+++ src/components/ResumePreview.jsx
@@ -5,11 +5,15 @@ import SummaryPreviewSection from "./SummaryPreviewSection";
 import ExperiencesPreviewSection from "./ExperiencesPreviewSection";
 import EducationPreviewSection from "./EducationPreviewSection";
 import SkillsPreviewSection from "./SkillsPreviewSection";
+import { useResumeStore } from "@/stores/useResumeStore";
 
 const { cn } = require("@/lib/utils");
 
-const ResumePreview = ({ classname }) => {
+const ResumePreview = ({ classname, resumeData, contentRef = null }) => {
   const containerRef = useRef(null);
+  const { resumes, currentResumeId } = useResumeStore();
+
+  const data = resumeData || resumes[currentResumeId] || {};
 
   const { width } = useDimensions(containerRef);
   return (
@@ -25,12 +29,31 @@ const ResumePreview = ({ classname }) => {
         style={{
           zoom: (1 / 794) * width,
         }}
+        ref={contentRef}
+        id="resumePreviewContent"
       >
-        <PersonalInfoPreviewSection />
-        <SummaryPreviewSection />
-        <ExperiencesPreviewSection />
-        <EducationPreviewSection />
-        <SkillsPreviewSection />
+        <PersonalInfoPreviewSection
+          personalInfo={data.personalInfo}
+          appearance={data.appearance}
+        />
+        <SummaryPreviewSection
+          summary={data.summary}
+          appearance={data.appearance}
+        />
+        <ExperiencesPreviewSection
+          workExperience={data.workExperience}
+          project={data.project}
+          appearance={data.appearance}
+          hasWorkExperience={data.hasWorkExperience}
+        />
+        <EducationPreviewSection
+          education={data.education}
+          appearance={data.appearance}
+        />
+        <SkillsPreviewSection
+          skills={data.skills}
+          appearance={data.appearance}
+        />
       </div>
     </div>
   );
